Data structure refactor.
We now have the possibility of race condition between reading and writing, since we're operating on buffers. To avoid having to place locks around reads, entities are now separate from the main cache key, with a unique identifer. The main cache key points to the current entity. On read, the entity ID is looked up first and remenbered througout the request so that concurrent writes cannot break anything.

This will require a garbage collection process.