Feature to blacklist cookies from esi evaluation
Cookies will still always be sent to subrequests for ESI includes
(unless disabled in another way), but any cookies blacklisted will not
evaluated in <esi:vars> or inside other esi tags, to prevent accidental
exposure of sensitive data.

This required a little refactoring, which will result in a small
performance increase on pages with lots of ESI cookie expansion.